requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
supabase==2.3.6
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
razorpay==1.4.1
Pillow==10.2.0
python-dotenv==1.0.1
google-generativeai
opencv-python-headless
httpx
